Add support for repeated groups
Add preliminary support for repeated groups, attempting to mimic the
structure of downloaded excel data. Stop trying to skip import of rows
that are already in the spreadsheet, as the code didn't handle updating
existing spreadsheet rows, which is even more important in the face of
repeated groups. Instead, always append survey data at the bottom of
each sheet.